compile under jdk1.2.2
git-svn-id: https://svn.apache.org/repos/asf/jakarta/poi/trunk@352677 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/Area3DPtg.java
@@ -68,7 +68,7 @@ import org.apache.poi.hssf.util.RangeAddress;
 
 public class Area3DPtg extends Ptg
 {
-    public final static short sid  = 0x3b;
+    public final static byte sid  = 0x3b;
     private final static int  SIZE = 11; // 10 + 1 for Ptg
     private short             field_1_index_extern_sheet;
     private short             field_2_first_row;
